Cleaned up some code / fixed PUT req
