+++ src/main/resources/application.properties
@@ -10,6 +10,7 @@ spring.datasource.username=sa
 spring.datasource.password=
 spring.datasource.driver-class-name=org.h2.Driver
 
+spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
 spring.jpa.hibernate.ddl-auto=create
 #spring.jpa.hibernate.ddl-auto=update
